refactor: extract ReaderPreferences and ReaderCSSGenerator from FolioReader
