+++ Sources/FolioReaderKit/FolioReaderKit.swift
@@ -137,7 +137,17 @@ public enum NavigationMenuBookListStyle: Int, CaseIterable {
 /// Main Library class with some useful constants and methods
 public class FolioReader: NSObject {
 
+    public static let FontSizes = ["15.5px", "17px", "18.5px", "20px", "22px", "24px", "26px", "28px", "30.5px", "33px", "35.5px"]
+    public static let DefaultFontSize = FolioReader.FontSizes[3]
+    public static let DefaultFontWeight = "500"
+    public static let DefaultLetterSpacing = 2
+    public static let DefaultLineHeight = 3
+    public static let DefaultTextIndent = 2
+
     public override init() { }
+
+    public lazy var preferences = ReaderPreferences(folioReader: self)
+    lazy var cssGenerator = ReaderCSSGenerator(folioReader: self)
 
     deinit {
         removeObservers()
@@ -218,436 +228,196 @@ extension FolioReader {
 
 extension FolioReader {
 
-    var preferenceProvider: FolioReaderPreferenceProvider? {
-        return delegate?.folioReaderPreferenceProvider?(self)
-    }
-    
-    func pref(boolFor key: ReaderPreferenceKey, default defaultValue: Bool) -> Bool {
-        return preferenceProvider?.preference(boolFor: key.rawKey, default: defaultValue) ?? defaultValue
-    }
-    
-    func pref(intFor key: ReaderPreferenceKey, default defaultValue: Int) -> Int {
-        return preferenceProvider?.preference(intFor: key.rawKey, default: defaultValue) ?? defaultValue
-    }
-    
-    func pref(stringFor key: ReaderPreferenceKey, default defaultValue: String) -> String {
-        return preferenceProvider?.preference(stringFor: key.rawKey, default: defaultValue) ?? defaultValue
-    }
-    
-    func pref(setBool value: Bool, for key: ReaderPreferenceKey) {
-        preferenceProvider?.preference(setBool: value, for: key.rawKey)
-    }
-    
-    func pref(setInt value: Int, for key: ReaderPreferenceKey) {
-        preferenceProvider?.preference(setInt: value, for: key.rawKey)
-    }
-    
-    func pref(setString value: String, for key: ReaderPreferenceKey) {
-        preferenceProvider?.preference(setString: value, for: key.rawKey)
-    }
-
-    /// Check if current theme is Night mode
+    @available(*, deprecated, message: "Use preferences instead")
     public var nightMode: Bool {
-        get {
-            pref(boolFor: .nightMode, default: false)
-        }
-        set (value) {
-            pref(setBool: value, for: .nightMode)
-
-            if let readerCenter = self.readerCenter {
-                UIView.animate(withDuration: 0.6, animations: {
-                    // _ = readerCenter.currentPage?.webView?.js("nightMode(\(self.nightMode))")
-                    readerCenter.pageIndicatorView?.reloadColors()
-                    readerCenter.configureNavBar()
-                    readerCenter.scrollScrubber?.reloadColors()
-                    readerCenter.collectionView.backgroundColor = (self.nightMode == true ? self.readerContainer?.readerConfig.nightModeBackground : UIColor.white)
-                }, completion: { (finished: Bool) in
-                    NotificationCenter.default.post(name: Notification.Name(rawValue: "needRefreshPageMode"), object: nil)
-                })
-            }
-        }
+        get { return preferences.nightMode }
+        set { preferences.nightMode = newValue }
     }
     
+    @available(*, deprecated, message: "Use preferences instead")
     public var themeMode: Int {
-        get {
-            pref(intFor: .themeMode, default: 1)
-        }
-        set (value) {
-            pref(setInt: value, for: .themeMode)
-            
-            guard let readerCenter = self.readerCenter,
-                  let backgroundColor = self.readerConfig?.themeModeBackground[self.themeMode] else { return }
-            
-            UIView.transition(
-                with: readerCenter.menuBarController.tabBar,
-                duration: 0.6,
-                options: .beginFromCurrentState.union(.transitionCrossDissolve),
-                animations: { () -> Void in
-                    readerCenter.menuBarController.tabBar.barTintColor = backgroundColor
-                },
-                completion: nil
-            )
-            
-            readerCenter.menuTabs.forEach { menu in
-                UIView.transition(
-                    with: menu.view,
-                    duration: 0.6,
-                    options: .beginFromCurrentState.union(.transitionCrossDissolve),
-                    animations: { () -> Void in
-                        menu.reloadColors()
-                    },
-                    completion: nil
-                )
-            }
-            
-            UIView.animate(withDuration: 0.6, animations: {
-                _ = readerCenter.currentPage?.webView?.js("themeMode(\(self.themeMode))")
-                readerCenter.pageIndicatorView?.reloadColors()
-                readerCenter.configureNavBar()
-                readerCenter.scrollScrubber?.reloadColors()
-                readerCenter.navigationItem.titleView?.subviews.forEach {
-                    if let label = $0 as? UILabel {
-                        label.textColor = self.readerConfig?.themeModeTextColor[self.themeMode]
-                    }
-                }
-                
-                readerCenter.collectionView.backgroundColor = backgroundColor
-                
-                if let page = readerCenter.currentPage {
-                    page.panDeadZoneTop?.backgroundColor = backgroundColor
-                    page.panDeadZoneBot?.backgroundColor = backgroundColor
-                    page.panDeadZoneLeft?.backgroundColor = backgroundColor
-                    page.panDeadZoneRight?.backgroundColor = backgroundColor
-                }
-            }, completion: { (finished: Bool) in
-                NotificationCenter.default.post(name: Notification.Name(rawValue: "needRefreshPageMode"), object: nil)
-            })
-        }
+        get { return preferences.themeMode }
+        set { preferences.themeMode = newValue }
     }
 
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentFont: String {
-        get {
-            pref(stringFor: .currentFont, default: "Georgia")
-        }
-        set (fontFamilyName) {
-            pref(setString: fontFamilyName, for: .currentFont)
-            readerCenter?.currentPage?.updateRuntimStyle(delay: 0.4)
-        }
+        get { return preferences.currentFont }
+        set { preferences.currentFont = newValue }
     }
 
-    static let FontSizes = ["15.5px", "17px", "18.5px", "20px", "22px", "24px", "26px", "28px", "30.5px", "33px", "35.5px"]
-    public static let DefaultFontSize = FolioReader.FontSizes[3]
-    
-    /// Check current font size. Default .m
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentFontSize: String {
-        get {
-            pref(stringFor: .currentFontSize, default: FolioReader.DefaultFontSize)
-        }
-        set (fontSize) {
-            pref(setString: fontSize, for: .currentFontSize)
-            readerCenter?.currentPage?.updateRuntimStyle(delay: 0.4)
-        }
+        get { return preferences.currentFontSize }
+        set { preferences.currentFontSize = newValue }
     }
     
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentFontSizeOnly: Int {
-        return Int(Double(currentFontSize.replacingOccurrences(of: "px", with: "")) ?? 20)
+        return preferences.currentFontSizeOnly
     }
 
-    public static let DefaultFontWeight = "500"
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentFontWeight: String {
-        get {
-            pref(stringFor: .currentFontWeight, default: "500")
-        }
-        set (fontWeight) {
-            pref(setString: fontWeight, for: .currentFontWeight)
-            readerCenter?.currentPage?.updateRuntimStyle(delay: 0.4)
-        }
+        get { return preferences.currentFontWeight }
+        set { preferences.currentFontWeight = newValue }
     }
     
-    /// Check current audio rate, the speed of speech voice. Default 0
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentAudioRate: Int {
-        get {
-            pref(intFor: .currentAudioRate, default: 1)
-        }
-        set (value) {
-            pref(setInt: value, for: .currentAudioRate)
-        }
+        get { return preferences.currentAudioRate }
+        set { preferences.currentAudioRate = newValue }
     }
 
-    /// Check the current highlight style.Default 0
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentHighlightStyle: Int {
-        get {
-            pref(intFor: .currentHighlightStyle, default: FolioReaderHighlightStyle.yellow.rawValue)
-        }
-        set (value) {
-            pref(setInt: value, for: .currentHighlightStyle)
-        }
+        get { return preferences.currentHighlightStyle }
+        set { preferences.currentHighlightStyle = newValue }
     }
 
-    /// Check the current Media Overlay or TTS style
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentMediaOverlayStyle: MediaOverlayStyle {
-        get {
-            let rawValue = pref(intFor: .currentMediaOverlayStyle, default: MediaOverlayStyle.default.rawValue)
-            return MediaOverlayStyle(rawValue: rawValue) ?? .default
-        }
-        set (value) {
-            pref(setInt: value.rawValue, for: .currentMediaOverlayStyle)
-        }
+        get { return preferences.currentMediaOverlayStyle }
+        set { preferences.currentMediaOverlayStyle = newValue }
     }
 
+    @available(*, deprecated, message: "Use preferences instead")
     public var defaultScrollDirection: FolioReaderScrollDirection {
-        self.readerContainer?.book.spine.isRtl == true ? .horitonzalWithPagedContent : .horizontalWithScrollContent
+        return preferences.defaultScrollDirection
     }
-    /// Check the current scroll direction. Default .defaultVertical
+
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentScrollDirection: Int {
-        get {
-            pref(intFor: .currentScrollDirection, default: defaultScrollDirection.rawValue)
-        }
-        set (value) {
-            pref(setInt: value, for: .currentScrollDirection)
-
-            let direction = FolioReaderScrollDirection(rawValue: currentScrollDirection) ?? defaultScrollDirection
-            readerCenter?.currentPage?.setScrollDirection(direction)
-        }
+        get { return preferences.currentScrollDirection }
+        set { preferences.currentScrollDirection = newValue }
     }
 
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentNavigationMenuIndex: Int {
-        get {
-            pref(intFor: .currentNavigationMenuIndex, default: 0)
-        }
-        set (value) {
-            pref(setInt: value, for: .currentNavigationMenuIndex)
-        }
+        get { return preferences.currentNavigationMenuIndex }
+        set { preferences.currentNavigationMenuIndex = newValue }
     }
     
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentAnnotationMenuIndex: Int {
-        get {
-            pref(intFor: .currentAnnotationMenuIndex, default: 0)
-        }
-        set (value) {
-            pref(setInt: value, for: .currentAnnotationMenuIndex)
-        }
+        get { return preferences.currentAnnotationMenuIndex }
+        set { preferences.currentAnnotationMenuIndex = newValue }
     }
     
-    /**
-     0: Grid
-     1: List
-     */
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentNavigationMenuBookListStyle: NavigationMenuBookListStyle {
-        get {
-            guard self.structuralStyle == .bundle else {
-                return .List
-            }
-            let defaults: NavigationMenuBookListStyle = self.structuralTrackingTocLevel == .level1 ? .Grid : .List
-            let rawValue = pref(intFor: .currentNavigationMenuBookListStyle, default: defaults.rawValue)
-            return NavigationMenuBookListStyle(rawValue: rawValue) ?? defaults
-        }
-        set (value) {
-            pref(setInt: value.rawValue, for: .currentNavigationMenuBookListStyle)
-        }
+        get { return preferences.currentNavigationMenuBookListStyle }
+        set { preferences.currentNavigationMenuBookListStyle = newValue }
     }
     
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentVMarginLinked: Bool {
-        get {
-            pref(boolFor: .currentVMarginLinked, default: true)
-        }
-        set (value) {
-            pref(setBool: value, for: .currentVMarginLinked)
-        }
+        get { return preferences.currentVMarginLinked }
+        set { preferences.currentVMarginLinked = newValue }
     }
     
+    @available(*, deprecated, message: "Use preferences instead")
     public var defaultMarginTop: Int {
-        (self.readerCenter?.traitCollection ?? UIScreen.main.traitCollection).verticalSizeClass == .regular ? 10 : 5    //5% for regular size, otherwise 2.5%
+        return preferences.defaultMarginTop
     }
+
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentMarginTop: Int {
-        get {
-            let defaults = self.defaultMarginTop
-            return pref(intFor: .currentMarginTop, default: defaults)
-        }
-        set (value) {
-            let newValue = max(0, min(50, value))
-            pref(setInt: newValue, for: .currentMarginTop)
-            guard currentVMarginLinked == false else { return }
-            readerCenter?.currentPage?.byWritingMode(
-                horizontal: { self.readerCenter?.currentPage?.updateViewerLayout(delay: 0.2) },
-                vertical: { self.readerCenter?.currentPage?.updateRuntimStyle(delay: 0.4) }
-            )
-        }
+        get { return preferences.currentMarginTop }
+        set { preferences.currentMarginTop = newValue }
     }
 
+    @available(*, deprecated, message: "Use preferences instead")
     public var defaultMarginBottom: Int {
-        (self.readerCenter?.traitCollection ?? UIScreen.main.traitCollection).verticalSizeClass == .regular ? 10 : 5    //5% for regular size, otherwise 2.5%
+        return preferences.defaultMarginBottom
     }
+
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentMarginBottom: Int {
-        get {
-            let defaults = defaultMarginBottom
-            return pref(intFor: .currentMarginBottom, default: defaults)
-        }
-        set (value) {
-            let newValue = max(0, min(50, value))
-            pref(setInt: newValue, for: .currentMarginBottom)
-            guard currentVMarginLinked == false else { return }
-            readerCenter?.currentPage?.byWritingMode(
-                horizontal: { self.readerCenter?.currentPage?.updateViewerLayout(delay: 0.2) },
-                vertical: { self.readerCenter?.currentPage?.updateRuntimStyle(delay: 0.4) }
-            )
-        }
+        get { return preferences.currentMarginBottom }
+        set { preferences.currentMarginBottom = newValue }
     }
 
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentHMarginLinked: Bool {
-        get {
-            pref(boolFor: .currentHMarginLinked, default: true)
-        }
-        set (value) {
-            pref(setBool: value, for: .currentHMarginLinked)
-        }
+        get { return preferences.currentHMarginLinked }
+        set { preferences.currentHMarginLinked = newValue }
     }
     
+    @available(*, deprecated, message: "Use preferences instead")
     public var defaultMarginLeft: Int {
-        (self.readerCenter?.traitCollection ?? UIScreen.main.traitCollection).horizontalSizeClass == .regular ? 30 : 5    //15% for regular size, otherwise 2.5%
+        return preferences.defaultMarginLeft
     }
+
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentMarginLeft: Int {
-        get {
-            let defaults = self.defaultMarginLeft
-            return pref(intFor: .currentMarginLeft, default: defaults)
-        }
-        set (value) {
-            let newValue = max(0, min(50, value))
-            pref(setInt: newValue, for: .currentMarginLeft)
-            guard currentHMarginLinked == false else { return }
-            readerCenter?.currentPage?.byWritingMode(
-                horizontal: { self.readerCenter?.currentPage?.updateRuntimStyle(delay: 0.4) },
-                vertical: { self.readerCenter?.currentPage?.updateViewerLayout(delay: 0.2) }
-            )
-        }
+        get { return preferences.currentMarginLeft }
+        set { preferences.currentMarginLeft = newValue }
     }
 
+    @available(*, deprecated, message: "Use preferences instead")
     public var defaultMarginRight: Int {
-        (self.readerCenter?.traitCollection ?? UIScreen.main.traitCollection).horizontalSizeClass == .regular ? 30 : 5     //15% for regular size, otherwise 2.5%
-    }
-    public var currentMarginRight: Int {
-        get {
-            let defaults = self.defaultMarginRight
-            return pref(intFor: .currentMarginRight, default: defaults)
-        }
-        set (value) {
-            let newValue = max(0, min(50, value))
-            pref(setInt: newValue, for: .currentMarginRight)
-            guard currentHMarginLinked == false else { return }
-            readerCenter?.currentPage?.byWritingMode(
-                horizontal: { self.readerCenter?.currentPage?.updateRuntimStyle(delay: 0.4) },
-                vertical: { self.readerCenter?.currentPage?.updateViewerLayout(delay: 0.2) }
-            )
-        }
-    }
-    
-    public static let DefaultLetterSpacing = 2
-    public var currentLetterSpacing: Int {
-        get {
-            pref(intFor: .currentLetterSpacing, default: 2)
-        }
-        set (value) {
-            pref(setInt: value, for: .currentLetterSpacing)
-            readerCenter?.currentPage?.updateRuntimStyle(delay: 0.4)
-        }
-    }
-    
-    public static let DefaultLineHeight = 3
-    public var currentLineHeight: Int {
-        get {
-            pref(intFor: .currentLineHeight, default: 3)
-        }
-        set (value) {
-            pref(setInt: value, for: .currentLineHeight)
-            readerCenter?.currentPage?.updateRuntimStyle(delay: 0.4)
-        }
+        return preferences.defaultMarginRight
     }
 
-    //in em
-    public static let DefaultTextIndent = 2
+    @available(*, deprecated, message: "Use preferences instead")
+    public var currentMarginRight: Int {
+        get { return preferences.currentMarginRight }
+        set { preferences.currentMarginRight = newValue }
+    }
+    
+    @available(*, deprecated, message: "Use preferences instead")
+    public var currentLetterSpacing: Int {
+        get { return preferences.currentLetterSpacing }
+        set { preferences.currentLetterSpacing = newValue }
+    }
+    
+    @available(*, deprecated, message: "Use preferences instead")
+    public var currentLineHeight: Int {
+        get { return preferences.currentLineHeight }
+        set { preferences.currentLineHeight = newValue }
+    }
+
+    @available(*, deprecated, message: "Use preferences instead")
     public var currentTextIndent: Int {
-        get {
-            pref(intFor: .currentTextIndent, default: 2)
-        }
-        set (value) {
-            pref(setInt: value, for: .currentTextIndent)
-            readerCenter?.currentPage?.updateRuntimStyle(delay: 0.4)
-        }
+        get { return preferences.currentTextIndent }
+        set { preferences.currentTextIndent = newValue }
     }
     
+    @available(*, deprecated, message: "Use preferences instead")
     public var doWrapPara: Bool {
-        get {
-            pref(boolFor: .doWrapPara, default: false)
-        }
-        set (value) {
-            pref(setBool: value, for: .doWrapPara)
-        }
+        get { return preferences.doWrapPara }
+        set { preferences.doWrapPara = newValue }
     }
     
+    @available(*, deprecated, message: "Use preferences instead")
     public var doClearClass: Bool {
-        get {
-            pref(boolFor: .doClearClass, default: true)
-        }
-        set (value) {
-            pref(setBool: value, for: .doClearClass)
-        }
+        get { return preferences.doClearClass }
+        set { preferences.doClearClass = newValue }
     }
     
+    @available(*, deprecated, message: "Use preferences instead")
     public var styleOverride: StyleOverrideTypes {
-        get {
-            let rawValue = pref(intFor: .styleOverride, default: StyleOverrideTypes.PNode.rawValue)
-            return StyleOverrideTypes(rawValue: rawValue) ?? .PNode
-        }
-        set (value) {
-            pref(setInt: value.rawValue, for: .styleOverride)
-            readerCenter?.currentPage?.updateRuntimStyle(delay: 0.2)
-        }
+        get { return preferences.styleOverride }
+        set { preferences.styleOverride = newValue }
     }
     
     @available(*, deprecated, message: "use delegate")
     @objc dynamic open var savedPositionForCurrentBook: FolioReaderReadPosition? {
-        get {
-            guard let bookId = self.readerCenter?.book.name?.deletingPathExtension else { return nil }
-            folioLogger("savedPositionForCurrentBook get")
-            return delegate?.folioReaderReadPositionProvider?(self).folioReaderReadPosition(self, bookId: bookId)
-        }
-        set {
-            guard let position = newValue,
-                  let bookId = self.readerCenter?.book.name?.deletingPathExtension else { return }
-            
-            guard self.isReaderReady || position.takePrecedence else { return }
-            
-            if let debug = readerConfig?.debug, debug.contains(.functionTrace) {
-                Thread.callStackSymbols.forEach {
-                    folioLogger($0)
-                }
-                if position.bookProgress < 5.0 {
-                    folioLogger(position.bookProgress.description)
-                }
-            }
-            
-            self.save(readPosition: position, for: bookId)
-        }
+        get { return preferences.savedPositionForCurrentBook }
+        set { preferences.savedPositionForCurrentBook = newValue }
     }
     
+    @available(*, deprecated, message: "Use preferences instead")
     public var structuralStyle: FolioReaderStructuralStyle {
-        get {
-            let rawValue = pref(intFor: .structuralStyle, default: FolioReaderStructuralStyle.atom.rawValue)
-            return FolioReaderStructuralStyle(rawValue: rawValue) ?? .atom
-        }
-        set {
-            pref(setInt: newValue.rawValue, for: .structuralStyle)
-        }
+        get { return preferences.structuralStyle }
+        set { preferences.structuralStyle = newValue }
     }
     
+    @available(*, deprecated, message: "Use preferences instead")
     public var structuralTrackingTocLevel: FolioReaderPositionTrackingStyle {
-        get {
-            let rawValue = pref(intFor: .structuralTrackingTocLevel, default: FolioReaderPositionTrackingStyle.linear.rawValue)
-            return FolioReaderPositionTrackingStyle(rawValue: rawValue) ?? .linear
-        }
-        set {
-            pref(setInt: newValue.rawValue, for: .structuralTrackingTocLevel)
-        }
+        get { return preferences.structuralTrackingTocLevel }
+        set { preferences.structuralTrackingTocLevel = newValue }
     }
 }
 
@@ -715,162 +485,28 @@ extension FolioReader {
 
 extension FolioReader {
     
-    
+    @available(*, deprecated, message: "Use cssGenerator instead")
     func generateRuntimeStyle() -> String {
-        let letterSpacing = Float(currentLetterSpacing * 2 * currentFontSizeOnly) / Float(100)
-        let lineHeight = Decimal((currentLineHeight + 10) * 5) / 100 + 1    //1.5 ~ 2.05
-        let textIndent = (letterSpacing + Float(currentFontSizeOnly)) * Float(currentTextIndent)
-        let marginTopEm = Decimal(1)
-        let marginBottonEm = lineHeight - 1
-        
-        
-        var style = ""
-        if styleOverride != .None {
-            var tagSelector = "p"
-            if styleOverride.rawValue >= StyleOverrideTypes.PlusTD.rawValue {
-                tagSelector += ", td"
-            }
-            if styleOverride.rawValue >= StyleOverrideTypes.PlusSPAN.rawValue {
-                tagSelector += ", td, span"
-            }
-            
-        style += """
-            \(tagSelector) {
-                /*font-family: \(currentFont) !important;*/
-                /*font-size: \(currentFontSize) !important;*/
-                /*font-weight: \(currentFontWeight) !important;*/
-                /*letter-spacing: \(letterSpacing)px !important;*/
-                /*line-height: \(lineHeight) !important;*/
-                /*text-indent: \(textIndent)px !important;*/
-                /*text-align: justify !important;*/
-                /*margin: \(marginTopEm)em 0 \(marginBottonEm)em 0 !important;*/
-                /*-webkit-hyphens: auto !important;*/
-            }
-            
-            span {
-                /*letter-spacing: \(letterSpacing)px !important;*/
-                /*line-height: \(lineHeight) !important;*/
-            }
-            
-            """
-        }
-        if let pageWidth = readerCenter?.pageWidth/*, let pageHeight = readerCenter?.pageHeight*/ {
-            let marginTop = 0 //CGFloat(currentMarginTop) / 200 * pageHeight
-            let marginBottom = 0 //CGFloat(currentMarginBottom) / 200 * pageHeight
-            let marginLeft = CGFloat(currentMarginLeft) / 200 * pageWidth
-            let marginRight = CGFloat(currentMarginRight) / 200 * pageWidth
-            
-            style += """
-            
-            /*body {
-                padding: \(marginTop)px \(marginRight)px \(marginBottom)px \(marginLeft)px !important;
-                overflow: hidden !important;
-            }
-            
-            @page {
-                margin: \(marginTop)px \(marginRight)px \(marginBottom)px \(marginLeft)px !important;
-            }*/
-            
-            """
-        }
-        
-        
-        return style
+        return cssGenerator.generateRuntimeStyle()
     }
     
-    static let CssLevelTags : [StyleOverrideTypes: String] = [.PNode: "p", .PlusTD: "td", .PlusSPAN: "span", .AllText: ""]
-    static func CssLevels(type: String, def: String) -> [String] {
-        CssLevelTags.map {
-            let separator = $1.isEmpty ? "" : " "
-            return "html body.folioStyleL\($0.rawValue)\(type) \($1), body.folioStyleL\($0.rawValue)\(type)\(separator)\($1) { \(def) }"
-        }.sorted()
-    }
-    
-    static func CssImgLevels(type: String, def: String) -> [String] {
-        CssLevelTags.map {
-            let separator = $1.isEmpty ? "" : " "
-            return "html body.folioStyleL\($0.rawValue)\(type) \($1) img.folioImg, body.folioStyleL\($0.rawValue)\(type)\(separator)\($1) img.folioImg { \(def) }"
-        }.sorted()
-    }
-    
+    @available(*, deprecated, message: "Use cssGenerator instead")
     func cssFontFamilies() -> String {
-        UIFont.familyNames.map {
-            FolioReader.CssLevels(type: "FontFamily\($0.replacingOccurrences(of: " ", with: "_"))", def: "font-family: \"\($0)\" !important;")
-        }.flatMap { $0 }.joined(separator: "\n")
+        return cssGenerator.cssFontFamilies()
     }
     
+    @available(*, deprecated, message: "Use cssGenerator instead")
     func cssUserFontFaces() -> String {
-        guard let readerConfig = readerConfig else { return "" }
-        
-        return readerConfig.userFontDescriptors.compactMap { fontName, fontDescriptor -> String? in
-//                let ctFont = CTFontCreateWithName(fontName as CFString, CGFloat(currentFontSizeOnly), nil)
-//                let ctFontSymbolicTrait = CTFontGetSymbolicTraits(ctFont)
-//                let ctFontTraits = CTFontCopyTraits(ctFont)
-//                let ctFontURL = unsafeBitCast(CTFontDescriptorCopyAttribute(fontDescriptor, kCTFontURLAttribute), to: CFURL.self)
-            guard let ctFontURL = CTFontDescriptorCopyAttribute(fontDescriptor, kCTFontURLAttribute),
-                  CFGetTypeID(ctFontURL) == CFURLGetTypeID(),
-                  let fontURL = ctFontURL as? URL else {
-                      return nil
-                  }
-            
-            guard let ctFontFamilyName = CTFontDescriptorCopyAttribute(fontDescriptor, kCTFontFamilyNameAttribute),
-                  CFGetTypeID(ctFontFamilyName) == CFStringGetTypeID(),
-                  let fontFamilyName = ctFontFamilyName as? String else {
-                      return nil
-                  }
-            
-            var isItalic = false
-            var isBold = false
-            
-            var cssFontWeight = "normal"
-            
-            if let ctFontTraits = CTFontDescriptorCopyAttribute(fontDescriptor, kCTFontTraitsAttribute), CFGetTypeID(ctFontTraits) == CFDictionaryGetTypeID() {
-                if let ctFontSymbolicTrait = CFDictionaryGetValue(
-                    (ctFontTraits as! CFDictionary),
-                    unsafeBitCast(kCTFontSymbolicTrait, to: UnsafeRawPointer.self))  {
-                    
-                    var symTraitVal = UInt32()
-                    CFNumberGetValue(unsafeBitCast(ctFontSymbolicTrait, to: CFNumber.self), CFNumberType.intType, &symTraitVal)
-                    
-                    isItalic = symTraitVal & CTFontSymbolicTraits.traitItalic.rawValue > 0
-                    isBold = symTraitVal & CTFontSymbolicTraits.traitBold.rawValue > 0
-                    
-                    cssFontWeight = isBold ? "bold" : "normal"
-                }
-//                let isItalic = ctFontSymbolicTrait.contains(.traitItalic)
-//                let isBold = ctFontSymbolicTrait.contains(.traitBold)
-                
-                
-                if let weightRef = CFDictionaryGetValue(
-                    (ctFontTraits as! CFDictionary),
-                    unsafeBitCast(kCTFontWeightTrait, to: UnsafeRawPointer.self)) {
-                    
-                    var weightValue = Float()
-                    CFNumberGetValue(unsafeBitCast(weightRef, to: CFNumber.self), CFNumberType.floatType, &weightValue)
-                    if weightValue < -0.49 {
-                        cssFontWeight = "100"   //thin
-                    } else if weightValue < -0.29 {
-                        cssFontWeight = "200"   //extralight
-                    } else if weightValue < -0.19 {
-                        cssFontWeight = "300"   //light
-                    } else if weightValue < 0.01 {
-                        cssFontWeight = "400"   //normal
-                    } else if weightValue < 0.21 {
-                        cssFontWeight = "500"   //medium
-                    } else if weightValue < 0.31 {
-                        cssFontWeight = "600"   //semibold
-                    } else if weightValue < 0.41 {
-                        cssFontWeight = "700"   //bold
-                    } else if weightValue < 0.61 {
-                        cssFontWeight = "800"   //extrabold
-                    } else {
-                        cssFontWeight = "900"   //heavy
-                    }
-                }
-            }
-            
-            return "@font-face { font-family: \"\(fontFamilyName)\"; font-style: \(isItalic ? "italic" : "normal"); font-weight: \(cssFontWeight); src: url(\"/_fonts/\(fontURL.lastPathComponent)\");} "
-            
-        }.joined(separator: " ")
+        return cssGenerator.cssUserFontFaces()
+    }
+    
+    @available(*, deprecated, message: "Use ReaderCSSGenerator.CssLevels instead")
+    public static func CssLevels(type: String, def: String) -> [String] {
+        return ReaderCSSGenerator.CssLevels(type: type, def: def)
+    }
+
+    @available(*, deprecated, message: "Use ReaderCSSGenerator.CssImgLevels instead")
+    public static func CssImgLevels(type: String, def: String) -> [String] {
+        return ReaderCSSGenerator.CssImgLevels(type: type, def: def)
     }
 }
